Migrate timestamps to uint64, add continuous auto mode with slew rate, and enhance dashboard interactivity
Firmware: replace uint32_t timestamps with uint64_t across history, API server, and SEN55 to
fix Y2038. Fan auto mode now uses continuous linear mapping (PM2.5 5–35 µg/m³ → 25–100%) with
2%/s slew rate instead of fixed thresholds. Partition layout updated for 40-byte samples.

Dashboard: update binary decoders for 64-bit timestamps and new header size. Replace stepped
auto mode display with gradient bar and ramping indicator. Add clickable fan metrics (speed,
RPM, PQ curve, auto mode) with info modals. Air column simulation gains fullscreen support
and physics scaling for canvas size. Fix "seconds ago" layout shift with opacity transition.

Collector: update binary decoder for new wire format, skip invalid readings (pm2_5 < 0).

+++ firmware/main/history.h
@@ -24,10 +24,10 @@
 #include "esp_err.h"
 
 /**
- * @brief Single data sample (36 bytes)
+ * @brief Single data sample (40 bytes)
  */
 typedef struct {
-    uint32_t timestamp;     // Unix timestamp (seconds since boot or epoch)
+    uint64_t timestamp;     // Unix timestamp (seconds since boot or epoch)
     float pm1_0;            // PM1.0 [µg/m³]
     float pm2_5;            // PM2.5 [µg/m³]
     float pm4_0;            // PM4.0 [µg/m³]
@@ -92,7 +92,7 @@ esp_err_t history_init(void);
  *
  * @return Timestamp in seconds
  */
-uint32_t history_get_timestamp(void);
+uint64_t history_get_timestamp(void);
 
 /**
  * @brief Record a new sample
@@ -155,7 +155,7 @@ void history_get_stats(history_stats_t *stats);
  * @param since_ts Timestamp threshold (returns samples with ts > since_ts)
  * @return Start index of first sample newer than since_ts (0 = all samples)
  */
-uint32_t history_get_samples_since(history_tier_t tier, uint32_t since_ts);
+uint32_t history_get_samples_since(history_tier_t tier, uint64_t since_ts);
 
 /**
  * @brief Clear all history data
